+++ cpp/subprojects/common/include/common/stopping/stopping_criterion.hpp
@@ -20,7 +20,7 @@ class IStoppingCriterion {
          * An enum that specifies all possible actions that may be executed, based on the result that is returned by a
          * stopping criterion.
          */
-        enum Action : uint32 {
+        enum Action : uint8 {
             CONTINUE = 0,
             STORE_STOP = 1,
             FORCE_STOP = 2
